ramips: mediatek: Turn off flow control for port 5 and 6 and globally PHY ports
This is a correction on my previous patch, as i actually enabled flow control on port 5 and not disabled it, whoops
Added seconds to sleep and allow the change to catchup.
MT7621 has a bug if flow control is enabled causing transmit timeout issue, so hopefully this fixes it now if as flow control wont come regardless of pc or switch.

Signed-off-by: David Bentham's avatarDavid Bentham <db260179@gmail.com>